added delete all dialogue

+++ pickapath/Main.java
@@ -312,6 +312,8 @@ public class Main extends JFrame {
 
 			@Override
 			public void actionPerformed(ActionEvent arg0) {
+				if (JOptionPane.showConfirmDialog(Main.this, "Are you sure you want to delete all?", "Delete?",
+						JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION)
 				canvas.deleteAllBoxes();
 
 			}
